launcher: don't show Dropbox folder when autostarted
When the user has arranged for com.dropbox.Client.desktop to be
symlinked into ~/.config/autostart (by hand or with Tweaks), they
probably want the sync service to start when they log in, but they
almost certainly don't want the Dropbox folder to pop up when they log
in.

gnome-session sets $DESKTOP_AUTOSTART_ID when launching .desktop files
from ~/.config/autostart and other autostart locations. Detect this
environment variable, and don't display the Dropbox folder if it is set.
When launched normally from the desktop, continue to display the Dropbox
folder (whether or not the daemon was already running).

Fixes #9.