Only launch file browser if daemon already running
Previously, launching the Flatpak would open the Dropbox folder in a
file browser, except if autostarted by gnome-session.

The upstream behaviour is to just open a tray icon. The rationale for
opening the file browser was to avoid user confusion: why would clicking
the Dropbox icon not show any visible feedback, except an easily-missed
16×16 tray icon? There are also systems that don't support tray icons at
all.

However, as seen by #8 and #93, autostart being noisy on non-GNOME
desktops is annoying. There does not seem to be a straightforward way to
detect whether one is being autostarted on other desktops (or at least
no-one has pointed to one).

As a compromise, only open a file browser if the Dropbox daemon was
already running. This leads to the following behaviours:

- If the .desktop file is symlinked into `~/.config/autostart`:
  - login: no file browser, tray icon appears if supported
  - click launcher: file browser (assuming Dropbox didn't quit in the
    meantime)
- Otherwise:
  - click launcher: no file browser, tray icon appears if supported
  - click launcher again: file browser (assuming Dropbox didn't quit in
    the meantime)