Update to new logo
Previously, the Dropbox logo was copied into this repo. There is now
a new logo, and nautilus-dropbox was updated in March 2020 to include
it.

Install the logo from nautilus-dropbox, and add
flatpak-external-data-checker configuration so it can be kept up to date
automatically in future.

Thanks to GitHub user pr0gr8mm3r for proposing updating the logo.